[253] 멀티시그 지갑과 연결된 싱글시그지갑 삭제 기능

+++ lib/providers/wallet_provider.dart
@@ -321,6 +321,9 @@ class WalletProvider extends ChangeNotifier {
   }
 
   Future<void> deleteWallet(int id) async {
+    final vaultIndex = _vaultList.indexWhere((element) => element.id == id);
+    if (vaultIndex == -1) return;
+
     if (await _walletRepository.deleteWallet(id)) {
       _setVaultList(_walletRepository.vaultList);
       await _preferenceProvider.removeVaultOrder(id);
